Move ExecutionStats into its own module
